Fixed: Breaking trakt API changes (Posters for add series are placeholders for now)

+++ src/NzbDrone.Core/Tv/Ratings.cs
@@ -5,7 +5,7 @@ namespace NzbDrone.Core.Tv
 {
     public class Ratings : IEmbeddedDocument
     {
-        public Int32 Percentage { get; set; }
+        public Decimal Percentage { get; set; }
         public Int32 Votes { get; set; }
         public Int32 Loved { get; set; }
         public Int32 Hated { get; set; }
